Isolate agency::bulk_then within its own header
