+++ agency/detail/bulk_functions/bulk_then_execution_policy.hpp
@@ -7,7 +7,9 @@
 #include <agency/detail/bulk_functions/decay_parameter.hpp>
 #include <agency/detail/bulk_functions/single_result.hpp>
 #include <agency/detail/bulk_functions/result_factory.hpp>
+#include <agency/detail/bulk_functions/bulk_invoke_execution_policy.hpp>
 #include <agency/detail/execution_policy_traits.hpp>
+#include <agency/detail/is_call_possible.hpp>
 #include <agency/execution_agent.hpp>
 #include <agency/executor_traits.hpp>
 #include <utility>
@@ -138,9 +140,9 @@ template<size_t... UserArgIndices, size_t... SharedArgIndices, class ExecutionPo
 bulk_then_execution_policy_result_t<
   ExecutionPolicy, Function, Future, Args...
 >
-  bulk_then_execution_policy_impl(index_sequence<UserArgIndices...>,
-                                  index_sequence<SharedArgIndices...>,
-                                  ExecutionPolicy& policy, Function f, Future& fut, Args&&... args)
+  bulk_then_execution_policy(index_sequence<UserArgIndices...>,
+                             index_sequence<SharedArgIndices...>,
+                             ExecutionPolicy& policy, Function f, Future& fut, Args&&... args)
 {
   using agent_type = typename ExecutionPolicy::execution_agent_type;
   using agent_traits = execution_agent_traits<agent_type>;
@@ -215,49 +217,6 @@ template<class ExecutionPolicy, class Function, class Future, class... Args>
 using is_bulk_then_possible_via_execution_policy = typename is_bulk_then_possible_via_execution_policy_impl<ExecutionPolicy,Function,Future,Args...>::type;
 
 
-template<bool enable, class ExecutionPolicy, class Function, class Future, class... Args>
-struct enable_if_bulk_then_execution_policy_impl {};
-
-template<class ExecutionPolicy, class Function, class Future, class... Args>
-struct enable_if_bulk_then_execution_policy_impl<true, ExecutionPolicy, Function, Future, Args...>
-{
-  using type = bulk_then_execution_policy_result_t<ExecutionPolicy,Function,Future,Args...>;
-};
-
-template<class ExecutionPolicy, class Function, class Future, class... Args>
-struct enable_if_bulk_then_execution_policy
-  : enable_if_bulk_then_execution_policy_impl<
-      is_bulk_then_possible_via_execution_policy<decay_t<ExecutionPolicy>,Function,Future,Args...>::value,
-      decay_t<ExecutionPolicy>,
-      Function,
-      Future,
-      Args...
-    >
-{};
-
-
 } // end detail
-
-
-template<class ExecutionPolicy, class Function, class Future, class... Args>
-typename detail::enable_if_bulk_then_execution_policy<
-  ExecutionPolicy, Function, Future, Args...
->::type
-  bulk_then(ExecutionPolicy&& policy, Function f, Future& fut, Args&&... args)
-{
-  using agent_traits = execution_agent_traits<typename std::decay<ExecutionPolicy>::type::execution_agent_type>;
-  const size_t num_shared_params_for_agent = detail::execution_depth<typename agent_traits::execution_category>::value;
-
-  return detail::bulk_then_execution_policy_impl(
-    detail::index_sequence_for<Args...>(),
-    detail::make_index_sequence<num_shared_params_for_agent>(),
-    policy,
-    f,
-    fut,
-    std::forward<Args>(args)...
-  );
-}
-
-
 } // end agency
