+++ testing/bulk_then/cuda/device_lambda/multiple_results.cu
@@ -1,5 +1,5 @@
+#include <agency/bulk_then.hpp>
 #include <agency/cuda/execution_policy.hpp>
-#include <agency/bulk_invoke.hpp>
 #include <iostream>
 
 template<class T, class Function>
